requirements.txt
langchain 
pinecone
openai 
fastapi 
uvicorn
langchain-community
tiktoken
langchain-openai
langchain-pinecone
python-multipart
jinja2
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
pydantic[email]
sqlalchemy-utils
